+++ History.rst
@@ -13,7 +13,7 @@
    limitations under the License.
 
 
-**Version 0.43.0 (17.03.2026)**
+**Version 0.43.0 (08.04.2026)**
 
     *Maintenance:*
 
@@ -23,7 +23,11 @@
     *Added:*
 
     * Possibility to exclude interface files from computation
-    * Markdown directives to support Python and **Robot Framework** code listings and log listings
+    * Markdown directives to support Python and **Robot Framework** code listings
+    * Markdown directives to support JSONP code listings
+    * Markdown directives to support common console and log listings
+    * Markdown directives to support listings of file system contents
+    * Support od **async** functions and methods
 
 **Version 0.42.0 (12.01.2026)**